Moving vendor to lib
